move final step to grapher and clean up famine names
