+++ etl/steps/data/grapher/wpf/2024-10-03/famines.py
@@ -1,5 +1,7 @@
 """Load a garden dataset and create a grapher dataset."""
 
+import pandas as pd
+
 from etl.helpers import PathFinder, create_dataset
 
 # Get paths and naming conventions for current step.
@@ -14,13 +16,24 @@ def run(dest_dir: str) -> None:
     ds_garden = paths.load_dataset("famines")
 
     # Read table from garden dataset.
-    tb = ds_garden["famines"]
+    tb = ds_garden["famines"].reset_index()
 
     #
     # Process data.
     #
-    # Rename for plotting.
-    tb = tb.rename_index_names({"conventional_title": "country", "date": "year"})
+    # Divide each row's 'wpf_authoritative_mortality_estimate' by the length of the corresponding 'Date' value to assume a uniform distribution of deaths over the period
+    tb["wpf_authoritative_mortality_estimate"] = tb.apply(
+        lambda row: row["wpf_authoritative_mortality_estimate"] / len(row["date"])
+        if pd.notna(row["date"])
+        else row["wpf_authoritative_mortality_estimate"],
+        axis=1,
+    )
+
+    # Unravel the 'date' column so that there is only one value per row. Years separated by commas are split into separate rows.
+    tb = unravel_dates(tb)
+    tb = tb.rename({"conventional_title": "country", "date": "year"}, axis=1)
+    tb = tb.format(["country", "year"])
+
     #
     # Save outputs.
     #
@@ -31,3 +44,13 @@ def run(dest_dir: str) -> None:
 
     # Save changes in the new grapher dataset.
     ds_grapher.save()
+
+
+def unravel_dates(tb):
+    """
+    Unravel the 'date' column so that there is only one value per row. Years separated by commas are split into separate rows.
+    """
+    # Split the 'date' column into multiple rows
+    tb = tb.assign(date=tb["date"].str.split(",")).explode("date").drop_duplicates().reset_index(drop=True)
+
+    return tb
